@@ -20,12 +20,16 @@ import com.amazonaws.services.sqs.AmazonSQS;
 import com.amazonaws.services.sqs.model.*;
 import org.opengroup.osdu.core.aws.sqs.AmazonSQSConfig;
 import org.opengroup.osdu.core.aws.ssm.K8sParameterNotFoundException;
+import org.opengroup.osdu.core.common.logging.DefaultLogger;
+import org.opengroup.osdu.core.common.logging.JaxRsDpsLog;
+import org.opengroup.osdu.core.common.model.http.DpsHeaders;
 
 import java.util.ArrayList;
 import java.util.List;
 import java.util.concurrent.ExecutionException;
 import java.util.concurrent.Executors;
 import java.util.concurrent.ThreadPoolExecutor;
+import java.util.concurrent.TimeoutException;
 import java.util.stream.Collectors;
 
 public class IndexerQueue {
@@ -40,6 +44,8 @@ public class IndexerQueue {
      * @throws InterruptedException
      */
 
+    private static final JaxRsDpsLog logger = LogProvider.getLogger();
+
     private static final int MAX_MESSAGE_ALLOWED = 100000;
     private static final int MAX_INDEX_THREADS = 50;
     private static final int MAX_BATCH_REQUEST_COUNT = 10;
@@ -50,11 +56,13 @@ public class IndexerQueue {
         String indexerUrl = String.format("%s/%s", environmentVariables.getTargetURL(), "api/indexer/v2/_dps/task-handlers/index-worker");
         String reIndexUrl = String.format("%s/%s", environmentVariables.getTargetURL(), "api/indexer/v2/reindex?force_clean=false");
         // System print lines go to cloudwatch automatically within ECS
-        System.out.println("Running Queue processor with the following environment variables:");
-        System.out.printf("Region: %s%n", environmentVariables.getRegion());
-        System.out.printf("Queue Name: %s%n", environmentVariables.getQueueUrl());
-        System.out.printf("Dead letter queue name: %s%n", environmentVariables.getDeadLetterQueueUrl());
-        System.out.printf("Target url: %s%n", environmentVariables.getTargetURL());
+        logger.info("Running Queue processor with the following environment variables:");
+        logger.info(String.format("Region: %s%n", environmentVariables.getRegion()));
+        logger.info(String.format("Queue Name: %s%n", environmentVariables.getQueueUrl()));
+        logger.info(String.format("Dead letter queue name: %s%n", environmentVariables.getDeadLetterQueueUrl()));
+        logger.info(String.format("Target url: %s%n", environmentVariables.getTargetURL()));
+        logger.info(String.format("Connecting to the SQS Queue: %s%n", environmentVariables.getQueueUrl()));
+
         while (true) {
 
             try {
@@ -63,14 +71,14 @@ public class IndexerQueue {
                 System.out.printf("Connecting to the SQS Queue: %s%n", environmentVariables.getQueueUrl());
                 AmazonSQSConfig sqsConfig = new AmazonSQSConfig(environmentVariables.getRegion());
                 AmazonSQS sqsClient = sqsConfig.AmazonSQS();
-                System.out.printf("Creating a thread pool with %s threads%n", MAX_INDEX_THREADS);
-                ThreadPoolExecutor executorPool = (ThreadPoolExecutor) Executors.newFixedThreadPool(MAX_INDEX_THREADS);
                 System.out.printf("Dead letter queue url: %s%n", environmentVariables.getDeadLetterQueueUrl());
                 messages = IndexerQueueService.getMessages(sqsClient, environmentVariables.getQueueUrl(), MAX_BATCH_REQUEST_COUNT, MAX_MESSAGE_ALLOWED);
                 System.out.printf("Processing %s messages from storage queue%n", messages.size());
 
 
                 if (!messages.isEmpty()) {
+                    System.out.printf("Creating a thread pool with %s threads%n", MAX_INDEX_THREADS);
+                    ThreadPoolExecutor executorPool = (ThreadPoolExecutor) Executors.newFixedThreadPool(MAX_INDEX_THREADS);
                     List<Message> reIndexMessages = messages.stream().filter(
                             msg -> msg.getMessageAttributes().containsKey("ReIndexCursor")
                     ).collect(Collectors.toList());
@@ -90,14 +98,15 @@ public class IndexerQueue {
             } catch (NullPointerException e) {
                 System.out.println(e.getMessage());
             } catch (Exception e) {
-                System.out.println(e.getMessage());
+                System.out.println("except: "+e.getMessage());
             }
 
         }
     }
 
-    public static void processIndexMessages(List<Message> messages, String indexerUrl, String queueUrl, String deadLetterQueueUrl, ThreadPoolExecutor executorPool, AmazonSQS sqsClient) throws ExecutionException, InterruptedException {
-        System.out.printf("Processing %s IndexMessage%n", messages.size());
+    public static void processIndexMessages(List<Message> messages, String indexerUrl, String queueUrl, String deadLetterQueueUrl, ThreadPoolExecutor executorPool, AmazonSQS sqsClient)
+            throws ExecutionException, InterruptedException, TimeoutException {
+        logger.info(String.format("Processing %s IndexMessage%n", messages.size()));
          /*
             Messages without authorization attributes will be removed them from queue and send directly to DLQ
          */
@@ -121,38 +130,38 @@ public class IndexerQueue {
          */
         // process only authorized messages
         List<IndexProcessor> indexProcessors = IndexerQueueService.processIndexQueue(authorizedMessages, indexerUrl, executorPool);
-        System.out.printf("%s Authorized Index Messages Processed%n", indexProcessors.size());
+        logger.info(String.format("%s Authorized Index Messages Processed%n", indexProcessors.size()));
 
         List<IndexProcessor> failedProcessors = indexProcessors.stream().filter(indexProcessor -> indexProcessor.getResult() == CallableResult.Fail || indexProcessor.getException() != null).collect(Collectors.toList());
-        System.out.printf("Total %s Index Messages Failed%n", failedProcessors.size());
+        logger.info(String.format("Total %s Index Messages Failed%n", failedProcessors.size()));
 
         // check if the message exponential visibility timeout
         IndexerQueueService.changeMessageVisibilityTimeout(sqsClient, queueUrl, failedProcessors);
         List<SendMessageResult> deadLetterResults = IndexerQueueService.sendMsgsToDeadLetterQueue(deadLetterQueueUrl, deadletters, sqsClient);
-        System.out.printf("%s Index Messages Dead Lettered%n", deadLetterResults.size());
+        logger.info(String.format("%s Index Messages Dead Lettered%n", deadLetterResults.size()));
         deleteMessageList.addAll(indexProcessors.stream().filter(indexProcessor -> indexProcessor.getResult() == CallableResult.Pass).collect(Collectors.toList()));
         List<DeleteMessageBatchRequestEntry> deleteEntries = deleteMessageList.stream().map(indexProcessor -> new DeleteMessageBatchRequestEntry(indexProcessor.getMessageId(), indexProcessor.getReceiptHandle())).collect(Collectors.toList());
-        System.out.printf("%s Index Messages Deleting (Succeeded and Non-Authorized)%n", deleteEntries.size());
+        logger.info(String.format("%s Index Messages Deleting (Succeeded and Non-Authorized)%n", deleteEntries.size()));
         List<DeleteMessageBatchRequest> deleteBatchRequests = IndexerQueueService.createMultipleBatchDeleteRequest(queueUrl, deleteEntries, MAX_BATCH_REQUEST_COUNT);
-        System.out.printf("%s Delete Batch Request Created%n", deleteBatchRequests.size());
+        logger.info(String.format("%s Delete Batch Request Created%n", deleteBatchRequests.size()));
 
         List<DeleteMessageBatchResult> deleteMessageBatchResults = IndexerQueueService.deleteMessages(deleteBatchRequests, sqsClient);
-        System.out.printf("%s Requests Deleted%n", deleteMessageBatchResults.size());
+        logger.info(String.format("%s Requests Deleted%n", deleteMessageBatchResults.size()));
     }
 
-    public static void processReIndexMessages(List<Message> reIndexMessages, String reIndexUrl, String queueUrl, ThreadPoolExecutor executorPool, AmazonSQS sqsClient) throws ExecutionException, InterruptedException {
+    public static void processReIndexMessages(List<Message> reIndexMessages, String reIndexUrl, String queueUrl, ThreadPoolExecutor executorPool, AmazonSQS sqsClient) throws ExecutionException, InterruptedException, TimeoutException {
         List<ReIndexProcessor> reIndexProcessors = IndexerQueueService.processReIndexQueue(reIndexMessages, reIndexUrl, executorPool);
         List<ReIndexProcessor> failedProcessors = reIndexProcessors.stream().filter(indexProcessor -> indexProcessor.getResult() == CallableResult.Fail || indexProcessor.getException() != null).collect(Collectors.toList());
-        System.out.printf("Total %s ReIndex Messages Failed%n", failedProcessors.size());
+        logger.info(String.format("Total %s ReIndex Messages Failed%n", failedProcessors.size()));
         reIndexProcessors.removeAll(failedProcessors);
         ArrayList<ReIndexProcessor> deleteMessageList = new ArrayList<>();
         deleteMessageList.addAll(reIndexProcessors.stream().filter(indexProcessor -> indexProcessor.getResult() == CallableResult.Pass).collect(Collectors.toList()));
         List<DeleteMessageBatchRequestEntry> deleteEntries = deleteMessageList.stream().map(indexProcessor -> new DeleteMessageBatchRequestEntry(indexProcessor.getMessageId(), indexProcessor.getReceiptHandle())).collect(Collectors.toList());
-        System.out.printf("%s ReIndex Messages Deleting (Succeeded and Non-Authorized)%n", deleteEntries.size());
+        logger.info(String.format("%s ReIndex Messages Deleting (Succeeded and Non-Authorized)%n", deleteEntries.size()));
         List<DeleteMessageBatchRequest> deleteBatchRequests = IndexerQueueService.createMultipleBatchDeleteRequest(queueUrl, deleteEntries, MAX_BATCH_REQUEST_COUNT);
-        System.out.printf("%s Delete Batch Request Created%n", deleteBatchRequests.size());
+        logger.info(String.format("%s Delete Batch Request Created%n", deleteBatchRequests.size()));
         List<DeleteMessageBatchResult> deleteMessageBatchResults = IndexerQueueService.deleteMessages(deleteBatchRequests, sqsClient);
-        System.out.printf("%s Requests Deleted%n", deleteMessageBatchResults.size());
+        logger.info(String.format("%s Requests Deleted%n", deleteMessageBatchResults.size()));
 
     }
 }
